+++ CHANGES.rst
@@ -16,7 +16,7 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 
 - Quote all components of a redirect URL (not only the path component)
   (`#1027 <https://github.com/zopefoundation/Zope/issues/1027>`_)
-  
+
 - Drop the convenience script generation from the buildout configuration
   in order to get rid of a lot of dependency version pins.
   These were only needed for maintainers who can install them manually.
@@ -27,6 +27,8 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 - Modify "manage_access" to allow users to switch from the compact view
   to the complete matrix view when more than 30 roles are defined.
   (`#1039 <https://github.com/zopefoundation/Zope/pull/1039>`_)
+
+- Allow leading ``.`` in cookie domain names.
 
 
 5.5.1 (2022-04-05)